Add the ability to allocate resources exclusively to a job
The trick is to add and "idle" tag to each core when the scheduler
first starts up.  As each core is allocated to a job, it's idle tag is
replaced by a lwj_id tag.  When the job completes, the lwj_id tag is
removed and the "idle" tag is put back.

When we are looking for resources for job, we invoke rdl_find() to
find resources that the job requires and that have an "idle" tag
attached.