rdl: Fixup changes introduced in commit 0466f1ed
A cherry-picked commit was inadvertently applied twice to rdl.c,
resulting in rdl objects being duplicated on the rdllib rdl_list.
This caused double-free of any rdl object and a crash during
rdl_destroy().

This commit cleans up the damage introduced by the stray commit.
Note: The commit could not simply be reverted since it was a dup,
the revert failed with conflicts.

Fixes #1