Alternate resource management implementation
This commit includes an alternate resource representation and new API
for the scheduler comms module.  This work is a precursor to storing
resources in the KVS and lays the groundwork to support multiple
graphs of resources.  All of the original functionality of the sched
module remains, which includes the dependence on wreck to launch
tasks.