resrc: Add resource tree support
Each resource record now contains a member that refers to the
resource's position in the physical resouce hierarchy. A job's request
for a resource composite is conveyed in the form of a json
object. Expanded the single test (tresrc) program to demonstrate the
creation of a resource composite request and the search to available matches
in a collection of hierarchical resources.