rdl: add function to set default tags for Resource types
A new class member '_defaulTags' may be set to define a list of tags
all resources of a given type (and its subclasses) are to inherit.
A class method `default_tags` is provided to set this class attribute.

For example:

 Resource:default_tags{ "idle" }

Will cause all subsequent resources to be set with an "idle" tag. Since
all resources types are a subclass of `Resource', this applies to all
resources in the file being evaluated after the time the default tags are
set.

To override the default tags for a Resource subtype, use the class
variable for that type, if it exists, e.g.

 Node.default_tags{ "idle", "compute" }

will additionally set a "compute" tag on all Node type resources.

The default tags of a subclass completely overrides the base class
definition, so default tags shared between these classes of resources must
be duplicated, as seen above.