resrc: depends on libuuid
libresrc uses functions from libuuid, but fails to link
with the library. The only reason this used to work is because
flux-core used to link against libuuid, but this is no longer
the case.

For now add the link to libuuid, but the code should probably
be converted to use zuuid* in the near term.