update flux-core API usage after changes
kvsitr_t  is now a struct not a pointer to struct.

kvs callback typedefs changed names.  Eliminate places in simulator
where this was being cast, as doing so is a bit dangerous - changes
in prototype would be hidden.  Case in point, these functions are
expected to return an integer result and were declared with return
type void.

Use the (deprecated) jsc _obj interfaces