resrc: Eliminate obsolete code
The original design included a hash table of all the resources.  The
thinking was that this table would provide an alternate means to
search and update resources that could perform better than a tree
search.  This facility has gone unused and now has serves no purpose
other than to unnecessarily complicate the code.