resrc: Add ability to read resources generated by hwloc
The sched module will now load the kvs's resource.hwloc resources when
the module is loaded without specifying the rdl-conf= option.
resrc_generate_xml_resources() was added to resrc.c to read a node's
hwloc resources in the xml format.

The resrc implementation remains independent of any flux instance.
I.e, it (still) does not rely on a handle to a flux instance.  Hence
the tresrc tests can be exercised without invoking flux.

tresrc was extended to include a test for
resrc_generate_xml_resource().  This test can only read the resources
of the node invoking the test.  So, the resource request construction
had to be simplified to eliminate multiple nodes and properties that
were only present in the sample RDL config file.

The sched module must make a separate call to
resrc_generate_xml_resource() for each node in the flux instance to
build up the complete resource tree.

This commit addresses the requirements documented in:
https://github.com/flux-framework/flux-sched/issues/61