sched:  extract the backfill scheduling code from the sched framework
A new backfillplugin1 was added as an alternative to the original FCFS
schedplugin1 plugin.

3 new API's were added to the sched framework: sched_loop_setup(),
allocate_resources(), and reserve_resources().

Includes many changes across sched as well as resrc.  resrc now
includes support for time-based as well as now-based resource
matching, allocations and reservations.  The backfill scheduler uses
the time-based functionality while the schedplugin1 FCFS uses the
now-based versions.

The stage_tree_list() was removed and instead staging was moved to the
selection function.  As size elements of a resource are selected, they
are also staged.

Finally, the t2002-easy.t test was modified to load the new plugin.
The hype-test.csv file of input jobs was not modified and the
easy_expected order remains the same.