rs2rank: sched targets the correct ranks to run a job on the allocated nodes
Fix the issue #69 where schedsrv incorrectly targets the first
k consecutive broker ranks (starting from rank 0) to launch a job
of k nodes, regardless of what nodes it is allocated to
and what broker ranks manage these nodes. For example, even if two
jobs are allocated back to back to cab[1-2] and cab[3-4], the old
scheme will end up launching both of the jobs into cab[1-2], assuming
rank 0 and 1 are running on these nodes.

This commit also abstracts out the multiple resource reader
option logic and resource-to-rank mapping logic into separate classes
to enhance code maintainability: i.e., rsreader and rs2rank.

Specifically, as part of the load_resources function, schedsrv
is modified to use rsreader's resource load API to populate
the resrc hierarchy and also use rs2rank APIs to link each
node-type resrc_t object in this hierarchy to its managing
broker rank(s).

The rs2rank table (of machs_t type) is the main data structure
to assist schedsrv with the linking. It is keyed by a hostname, and
each table entry maintains the signature of each per-rank
resource partition (e.g., core0-7) within the host
as well as one or more equivalent broker ranks that manage
the resource partition.

We identify each per-rank resource partition by taking the SHA
digest of per-rank hwloc xml data (e.g., resource.hwloc.xml.0
for rank 0). Thus, if two ranks on the same host happen
to have the identical xml data under this key, they are
treated as equivalent ranks.

resrc_t class is also modified such that each node-type
resrc object gets filled in with the per-rank signature during
the linking phase. Once linked, its hostname and signature
are used to uniquely identify the ranks set that can/should spawn
the job on the allocated nodes.

During the linking process, this commit also performs a consistency
check by comparing the built resrc hierarchy with the hwloc data
by using rs2rank as well.
If inconsistency is detected, it prints out an error message and
simply discard the built hierarchy and rebuild a new one
based on hwloc xml data.

Finally, for emulation mode (RSREADER_RESRC_EMUL), because
the rdl and hwloc xml data are almost always inconsistent,
we simply force a link by associating all of the node-type
resrc_t object with the rank 0. This can be easily extended
to support fake hwloc mode for scalability testing beyond
emulation.

However, note that this is a stopgap solution as it serially
establishes the link between each resrc_t object and its managing rank.
A better solution is to do this in a distributed way, which
should be implemented as part of flux-core.

Meanwhile, this commit also factors out the command line
argument parsing