resrc: Change resrc_match_resource() to match a request
The old argument to resrc_match_resource() matched a specified
resource against a sample resource.  The sample resource was the
resource part of a resource request object.  This proved to be
combersome as we proceed with the next round of changes.  So, now we
pass the whole request to resrc_match_resource().