resrc: Add the option to request an exclusive allocation
This commit addresses:
https://github.com/flux-framework/flux-sched/issues/101.  One change
is that resources will no longer be exclusively allocated to a job by
default.  Instead, the resource request includes the option for an
exclusive allocation and this will determine whether the resource is
dedicated exclusively to the job.

As commented in schedule_job() in schedsrv.c, requesting a size == 1
preserves the old behavior of only one job per resource - which the
simulator tests expect.  At some later point, I will add a simulator
test that exercises multiple jobs per resource, or size == 0.