rs2rank: add shareness test scripts
Add shareness tests that utilize a new hwloc front-end command
(`flux hwloc`) within `flux-core` to load in synthentic hwloc xml files
and submit and run a bunch of unit jobs and verify the
results (e.g., the correct rank spawned the job.)

The description of each test should be self-exploratory
by means of the name of the script and printing message of
each subtest.

In addition, use the absolute path to flux-waitjob
The latest flux-core commandline option change somehow
broke waitjob by not being able to invoke an external flux
command like this with -x option.

Providing the absolute path to flux-waitjob works
around the problem for now.