schedsrv: add verbosity flag to sched module
Add verbosity level switch, verbosity=<num>, to the load line of the sched module.
It is currently used to avoid printing of the resrc tree state all the time
with DEBUG_LOG. It can later be used to print debug or info
messages with different level of details for sched module.