testsuite: more efficient use of LONGTEST prereq
Set LONGTEST prereq in one place in sharness.d/sched-sharness.sh so that
it is available by default to all sharness tests in sched.

For t1003-stress.t and t1002-rs2rank-64ranks.t, all tests require LONGTEST,
so short circuit the entire suite when LONGTEST is not set. This avoids
unnecessarily starting a flux session when no tests will be run. (And in the
case of t1002-rs2rank-64ranks.t, it avoids starting a very large session
on a single node)