rdl: rename flux-cpuset module to flux.cpuset
Rename flux-cpuset.so to cpuset.so and put in a flux subdirectory of
LUA_CPATH so that cpuset module is loaded as `flux.cpuset`. A
convenience link is made from "builddir/rdl/flux/.libs/cpuset.so"
to "builddir/rdl/flux/cpuset.so" so that in-tree loading of the
module can occur the same way as installed path.

Update all users and ensure LUA_CPATH points to

  $(top_builddir)/rdl/?.so

instead of "rdl/.libs/?.so".

This required using FLUX_LUA_CPATH_PREPEND instead of LUA_CPATH
in t/Makefile.am to ensure we load the in-tree modules instead of
installed path modules when using installed 'fluxometer.lua'.

Fixes #131