testsuite: fix lua tests against installed sched
Only export FLUX_LUA_PATH_PREPEND and FLUX_LUA_CPATH_PREPEND when
FLUX_SCHED_TEST_INSTALLED is not set in the current environment.

We still set LUA_PATH correctly in order to find the fluxometer.lua
distributed with flux-core, but we do not want to prepend build and
sources paths to TESTS_ENVIRONMENT when we actually want to test
installed flux-sched components.

This is mainly for the fluxometer-based Lua tests. Other test env
is setup directly by sharness.