testsuite: fix non-POSIX warning from automake
Fix use of non-POSIX $(if ...) usage in t/Makefile.am to keep automake
from complaining. The replacement uses POSIX variable substitution to
substitute the build-tree LUA_PATH and LUA_CPATH to the PREPEND variables
iff FLUX_SCHED_TEST_INSTALLED *is not* set.

When FLUX_SCHED_TEST_INSTALLED *is* set, then the value of that variable
(`t` or maybe `1`) is substituted, which still isn't quite right, but at least
the build-tree Lua modules will not be used. Unfortunately, this was the
best that we could do at this time.

Fixes #136