resrc and sched:  Replace tree list searches with tree searches
Modifies the resrc search code to search a resource tree instead of a
resource tree list.  This has two benefits:

Can now search for an isolated resource (like cores) without regard to
its parent resource (like nodes).  Formally, we had to define a
request that include one node and multiple cores to find resource
composites of node+cores.  Now we can just search for cores.

The second benefit is that the code now conforms to the RFC 4
requirement:

"The job ID for a job that is allocated a resource in a composite
hierarchy MUST be annotated not only to the resource, but to each
parent up the tree of those resources allocated to the enclosing
instance"

The API's that change with this commit include:

* resrc_tree_search()

as well as four of the sched plugin's API's:

* find_resources()
* select_resources()
* allocate_resources()
* reserve_resources()

along with a corresponding change to the flux_lwj_t structure.

The tests in resrc/test/tresrc.c were also revised to match the new
API.

The rsreader library also needed mods to conform to the changes from
the tree_list to the tree search model.

The commit also includes a refactoring simplification to the
resrc_reqst's match_child() as well as the sched plugins'
select_child() calls that eliminates redundant code.