resrc_flow:  Introduce the resource flow module
Modeled after the resource tree module, the resource flow module provides
the means to represent a flow like power or network bandwidth.  Each
vertex in the flow graph has a capacity value.  Portions of the flow can
be allocated to jobs such that the total flow allocated to jobs cannot
exceed the capacity.