resrc:  Add resrc_unstage_resrc()
Modified the staging function to increment the staged value.  Formerly
the staged value was always set to the requested value.  Now a
resource will be able to respond to multiple staging requests and
increment the staged value each time.  This required the new
resrc_unstage_resrc() to do what resrc_stage_resrc(0) used to do.