resrc:  Introduce flow graph support.
Expanded the resrc module to support associations to hierarchies
beyond just the physical.  Multiple hierarchies are now represented by
the graphs member.

The existing RDL was leveraged to provide the configuration for the
additional hierarchies (power and bandwidth).  The configuration for
these two hierarchies was added to conf/hype.lua.

Ideally, the Node class should be reworked to not require socket
members.  Until this is revised, the power and bandwidth hierarchies
will extend into the socket and core resources.

The resrc_flow module implements the alternate hierarchies like power
and bandwidth.  The tresrc test has been extended to load and allocate
power and bandwidth.

Currently the "node" resource serves as the junction for the multiple
hierarchies.  The design will allow other resources to have their
respective hierarchies.

Yet to do is add the additional hierarchies to the sched module, but
the rudiments are in place.