travis-ci: run backtrace-all on failure
In order to support coredumps, install apport and gdb into
the travis-ci image. Installation of apport will allow
coredumps to be saved, since default travis container
core_pattern is piped to apport, and the core dumps are
lost when apport is not installed (and there is no current
way in travis to change the default core_pattern).

Apport will obey current corefile limit, so the corefile limit must
also increased to unlimited.

Finally, in after_failure script the backtrace-all.sh script is
always run to find and print backtrace for any corefiles left

(courtesy grondo in flux-core commit 66e5502)