jansson: work around lack of JSON_DECODE_ANY
Prior to v2.3.0, Jansson does not allow decode of values
that are not full json objects or arrays in json_loads().

Add a very kludgy workaround that attempts to wrap a value
in { "foo": <value> } to see if we can pull out the value
as a json_t through trickery (only if json_loads() fails).

This will likely be very slow and it might be better to just
require libjansson >= 2.3.