planner: replace twindow with planner-based
Replace resrc's twindow using planner-based API.
two things to note:

1. Because of an API mismatch between
resrc's "avail" functions and planner's, this commit
changes the signiture of the resrc's "avail" functions a bit
and adjust the resrc test cases.

2. Because there is yet no use case for resrc's copy
constructor, this commit doesn't make a deep copy of
the planner-based twindow member. The exact semantics
of the copy constructor needs to be discussed.