submit: add num-jobs argument to limit number of jobs in a sim
avoids the hangs we were experiencing when the number of jobs in the
input file exceeds the number of jobs passed in to the
intial_program. For example, if the user requests a simulation of 10K
jobs but passes in a file with 500K jobs, the initial_program will
terminate after 10K jobs complete but the broker will hang
indefinitely. This ensures that only N jobs ever make it into the
simulation in the first place.