simulator: [cleanup] rework get_job_from_kvs() helpers
As part of migrating off of deprecated KVS functions -

Rename helper functions that load values from the KVS
so they will not be confused with similar core functions.
Reduce the "typed" helper functions to two: one for
strings, and one for scalars.

Check function return values for error and reduce logging
to a single failure message if something goes wrong,
mirroring the error handling in put_job_in_kvs().