sched: Add priority plugin
The new priority plugin can be loaded in addition to the existing fcfs
or backfill plugins.

With the introduction of the priority plugin, the existing "sched_plugin"
had to be renamed to the "behavior_plugin".  "sched_plugin" now refers to
any plugin loaded by the sched module.  The two categories of sched
plugins, with different API's, are the behavior and priority plugins.