t/t1004-module-load.t: use kvs get -j for lwj status
Problem: The "module-load: no jobs are lost" subtest
of sharness t1004-module-load.t is failing.

The test is comparing "status" (with quotes) to status
(without quotes).  The value in the KVS is a JSON string,
but flux-framework/flux-core#1257 changed flux-kvs so
that it displays raw values by default, in this case the
encoded JSON string with the quotes.

Add -j to the flux kvs get command in the test, to get
the old behavior.