sched: Further development of the multi-factor priority plugin.
This is a continuation (but not a completion) of the multi-factor
priority plugin that was initiated by Don Lipari.  Don retired
before there was time to complete the plugin.

The code was uplifted from C to C++.  The idea behind that change
is that some of the core scheduler code is also C++, and a native
C++ priority plugin would allow an API where the parts can directly
exchange STL data structures such as vectors and remain RAII through
out the process rather than dropping down to C and manual memory
management.

The parser for the account/user information from the SLURM sacct
command had numerous bugs.  Many are now fixed.  Also, the parser
and the rest of the code were improved to allow the sacct information
to be reloaded and adjusted online (previously the data was completly
static after initial read).

Some of the sacctmgr parsing bugs were:

Add better error handling for missing "|".  A missing "|" would have
resulted in the an entry containing information from the previous line.
Now at least the bad line will be ignored with and error.

The code failed to properly copy temporary strings.  That is also resolved.
prio_node_create() makes copies of its inputs.  Proper error handling is
added for memory allocation failures, and prio_node_destroy() is added
to clean up the structure.

Also we correct the code to allow the "if (user)" check to function.
Previously "user" was a pointer to an empty string, meaning that
the condition would never be false.  Now the parser checks for an
empty field and sets "user" to NULL rather than pointing to an
empty string.  If white space were ever to appear in the field,
this check would not function correctly.

Also, often in the code the term "child" was used when it really
meant "sibling".  Naming was changed to better reflect the behavior.

A simple first test command is added.

norm_usage/norm_shares were caluclated incorrectly.  That has been
fixed.

The code comment describing the algorithm did not really match
the algorithm.  That has been somewhat improved.

Under the "sched" directory make a new "priority" sub-directory.
The "priority" direcotry will contain the various priority plugins.
The first priority plugin is in the mod_fair_tree (Modified Fair Tree)
subdirectory.

The files in the plugin are first compiled into a libtool convenience
library which is in turn used to both build the shared library and
directly linked against by test applications.  Test applications can
include either of the headers, depending on testing needs.  The shared
library will only export the functions that appear in
the "priority_external_api.hxx" file, so use of the internal header
and the convenience library give additional symbol access for testing.

The implementation of the fair share algorithm is improved to begin
genericizing it from some of the odd sacctmgr designs (e.g. the
"parent" account changes column from either "Parent" or "Account"
based on the values of _other_ columns).

The recursion is cleaned up considerably.

Make it possible to add and remove accounts/users (WIP).  While
accounts can be removed (just load in a new set of data from sacctmgr
that has the account missing), there is perhaps more work to be
done here.  For instance, if a user has a job running and meanwhile
that user is removed from the job's current account, what needs to
happen?  Presumably we need to keep record of that user until all
jobs from that user are complete so that resource usage can correctly
be applied to the parent accounts up the tree.  So the user
perhaps the user needs to be marked deleted and no longer eligible
for new jobs, but kept around until all current jobs are complete.

Duplicate accounts are ignored.

operator<< functions are provided for debugging purposes.