resource: Add Scoring API
Introduce data types and code that allows a matcher callback to
evaluate and score the visiting resource vertex.

A scoring API object passes the result of the subtree or up traversal
to the match callback methods so that they can use that information
to check the hiearchical constraints. (See the Resource Selection Policy
section of utilities/README.md).

Store the edge information along with evaluated scores so that
best-matching edges will be selected by the infrastructure layer.