resource: Add matcher_data_t type
Provide a match callback object (a derived class of matcher_data_t)
with base idioms to specify the target subsystems and resource
relationship types.

These then interface with the scheduler infrastructure to filter
the graph data accordingly as well as to determine the traversal
details.

Also implenment select_count () method that return the effective
count given the min and max of a resoure request and the number
of qualified resources evaluated.