resource-query: Add the initial implementation
Add a command-line utility that takes in an HPC resource
request written in Flux's Canonical Job Specification
(or simply a jobspec) (RFC 14) and selects the best-matching
compute and other resources in accordance with a selection policy.

Designed to serve as a debugging and testing interface
for our scheduling infrastructure

The main cli command is "match".

This command takes in a jobspec file name and either
allocates or reserves its best-matching resources.

Likewise, this command provides two subcommands: "allocate" will
try to allocate the best-matching resources
for the given jobspec; "allocate_orelse_reserve" will try to reserve
the resources into the future
(i.e., earliest possibly scheduable point), if an allocation
cannot be created on the current resource state.

The output format of an allocation or reservation is a reversed
tree shape where the root resource vertex appears at the last line.
This is less than ideal, but we put only a minimal effort
to the output format as the format will change in a near
future anyway: we will determine what "R" should be as
part of a co-design with flux-core's jobshell functionality.