test: Add resource-query tests for full vs. partial spec
Test various full and partial hierarchical specifications
using a medium sized machine configuration in
t/data/resource/grugs/medium.graphml
and jobspec files that reside in t/data/resource/jobspecs/*.

Test the correctness of resource-query for the fact that
the resource section of a jobspec can be fully
or partitially hierarchically specified. A fully specified
request describes the resource shape fully from the root
to the requested resources with respect to the resource
graph data used by resource-query.
A partially specified resource request omits the prefix
(i.e., from the root to the highest-level resources in
the request).

The text commands files are in
t/data/resource/commands/omit_prefix/* and they contain
a set of match commands on these jobspec files.

Print the best matching resources into an output file in the end,
and the output is compared against the expected output file
in t/data/resource/expected/omit_prefix.