test: Add resource-query tests for simple power scheduling
Add tests for matching power requirements using a machine
configuration in t/data/resource/grugs/power.graphml
and jobspec files t/data/resource/power/*.

power.graphml grug file specifies two subsystems. One
for containment and the other for power.

Use the Power-Aware (PA) selection policy (-S PA) of
resource-query to use the power subsystem as the
dominant subsystem to perform matches.

Limit test jobspecs to specify only down to the compute
node level as the fringes of the power subsystem
tree in power.graphml are at the compute node level
(associations are made between the PDU level in
the power subsystem and compute-node level in
the containment subsystem).

The text commands files are in
t/data/resource/commands/power/* and they contain
a set of match commands on these jobspec files.

Print the best matching resources into an output file
in the end, and the output is compared against the expected
output file in t/data/resource/expected/power.