sched: performance/scalability optimization for simple inorder FCFS
Avoid traversing the entire resource tree to release
reservations at every schedule loop invocation when a very simple,
high performance, in-order (queue-depth=1) FCFS policy is used.

Add get_sched_properties() into the scheduler policy plugin API
so that the scheduler framework comms module can get the properties
of the loaded policy scheduler plugin in order to turn on/off
this optimization.

Initially, only support whether the loaded sched plugin is
out-of-order scheduling capable. Backfill schedulers and FCFS
with queue-depth > 1 can always schedule jobs out of order.

Somewhat counterintuitive, but FCFS schedulers requires
the reservation capability for the general case.

If the first job doesn't find all of the required resources,
the FCFS scheduler reserves any partially found resources so that
it can move on to the next job to schedule.

The reason for this out of order behavior is that the next job
may only require resources of different type than what the first job
needs, and the fact that the first job is not scheduled
shouldn't prevent it from being scheduled -- that is,
without having to use the resources that the first job can use
at the next schedule loop.

But with queue-depth=1, this out-of-order behavior isn't
possible and we use this for optimization -- bypassing the
reservation release.

We can view this as part of scheduler specialization.