resrc: Add lightweight R emitting support
Add resrc_tree_serialize_lite as the main function to
traverse the hardware hierarchy to emit the resources
in the lightweight R format.

Add support so that the upper layer can pass in
the resources types to emit and how they and
their children (i.e., the next level resource
type to be emitted) should be emitted.

This way, the resrc layer can still remain resource
type agnostic and the upper layer passes in
the input that control emit behavior.