build: use AX_BOOST_GRAPH macro
Problem: --with-boost=path has no effect on the search
location for libboost_graph.so.

boost_graph is found with AC_CHECK_LIB, while the rest of boost
is found with purpose-build autoconf macros.  The macro for
boost_graph is included but not used.

Call AX_BOOST_GRAPH and adjust Makefiles accordingly.