resrc: Add node exclusive scheduling
Add node-exclusive=[true|false] as a sched load option.

Add the exclusivity check in the resrc_tree_search function
such that we don't lead to overscheduling under node-exclusive
scheduling mode.

The purpose of the else branch in this function, which is
the branch taken when the resource match test on the
visiting resource vertex fails, is to select the high-level
resources leading to each exclusively allocated resource.
This is needed when the job request is partially specified,
for example, when only 1 core is requested. It allows
for selecting the cluster, node, socket that contain
this particular core.

For node-exclusive scheduling, however, this branch
must check exclusivity so that we stop recursing down
the hardware tree if the node vertex is already allocated
to another job.