rs2rank: Fix overscheduling when running multiple ranks on a node
In hwloc reader mode under multiple ranks on a node,
the hwloc data reported from the these ranks are exactly identical.
In this configuration, rs2rank groups those ranks
as a equivalent set and round-robins across for them
for execution requests.

This is the correct semantics when we use rdl reader
since we use this hwloc data only to link the resrc data
with those equivalent ranks. However, when we use hwloc
reader only mode, it is incorrect. We should rather
treat each rank as a distinct resource set
to facilitate testing.

Fix this issue by introducing an auxilliary field
to the signature input.

Allow each reported, identical hwloc data to
generate a different signature using this field.