libutil: remove const qualifier on shortjansson Jtostr
Problem: The shortjansson.h version of Jtostr() uses json_dumps()
which returns memory that must be freed by the caller. The const
qualifier on this function therefore encourages memory leaks.
This was probably in place because the equivalent function in json-c
*did* return a const char *.

Remove the const qualifier from this function and update the comment.
Actually, it would probably be better to remove Jtostr() altogether,
but that is left for future work.