sched: fix memory leaks in use of Jtostr()
When shortjson.h was replaced with shortjansson.h, the result of
Jtostr() went from a const char * which didn't need to be freed
by the caller, to a malloc'd string that *does* need to be freed
by the caller.

Clean up all uses of Jtostr() and be sure to free the result.