planner: Limit planner to handle only one resource type
Solve a problem where the mintime resource tree within planner
does not work for multi-resource types.
The problem is described in Issue #319.

Change the API and implementation of the planner layer
so that it now only concerns only a single resource type.
With this change, the upper layer (resource-query etc) will
now have to use multiple planner objects if it needs to
track multiple resource types.

Note that we considered other alternatives: e.g.,
introducing kd tree or  making each scheduled point node in the
resource binary search tree to maintain sub-resource BST
for the next significant resource types. But these
alternatives will significantly increase the software
complexity for a single software layer, and we decide
to push the complexity to the upper layer.

Finally, adjust the unit test case as well.