resrc: Add hwloc-based GPU support
Change the way the resrc layer walks the hwloc hierarchy
from an array style to a recursive style. Needed to walk
the IO hierachy (see Issue #352).

Add support into resrc to recognize and use CoProc devices
as gpu resource type. Include CUDA GPU as well as
AMD GPU (at least in theory since I couldn't test this on AMD).

Parse the name of these os-device hwloc objects to get the
logical index to be used as the IDs of GPUs.
While one may argue this is not robust, this is just about
the only way to get the correct index without having
to use a local hack. (see Issue #352).