rdl: force liblua symbols RTLD_GLOBAL
Problem: When the sched module loads librdl, which in turn loads Lua
code which uses `require` to load a lua DSO, lua_* symbols aren't
available and we get a warning like:

  undefined symbol lua_gettop

This is probably because sched itself is dlopened and all symbols it
links do not have RTLD_GLOBAL visibility (this is a guess).

Fix the issue by dlopening LIBLUA_SO which is exported by configure
as the filename of the liblua DSO found at build time.