rdl: Lua 5.2 compatibility for librdl
librdl made use of the following assumptions/defines that are not
compatible with Lua 5.2

 - Each Lua "thread" (coroutine) has its own globals table
 - Use of special index LUA_GLOBALSINDEX

For Lua 5.2 compatibility, remove the use of lua_newthread() and instead
share the same Lua state across each rdllib_init() instantiation. Set a
function environment on each method or "dostring()" call to use a special
per-rdl shadow globals table, stored in the registry.

An rdl_lua_setfenv() is added which does the right thing to either adjust
the environment of a function in 5.1, or set the first upvalue in Lua 5.2.