configure.ac: define LIBLUA_SO
Problem: librld code needs to dlopen liblua DSO in order to force
Lua symbols to have global visibility. Currently, librdl tries a
heuristic to attempt to open the same liblua that we linked with
at build and runtime. This is a brittle solution.

Since AX_LUA_LIBS already finds the correct name for liblua.so,
define the name in config.h as LIBLUA_SO so that librdl can switch
to using dlopen (LIBLUA_SO, RTLD_GLOBAL).