jobspec: Copy libjobspec from flux-core
Pull in jobspec.cpp and jobspec.hpp. The goal is two-fold.

1. Tightly managing libjobspec from within flux-sched solves odd
make-check errors that we are seeing on some platforms (e.g., Sierra).
The issues seem to arise due to C++ compiler version mismatches.

2. flux-core will transition away from libjobspec anyway
in a near future, and doing this now will help flux-core
make this transition without having to worry about
any breakage in flux-sched.

Name the library as libjobspec_conv.la instead
of libjobspec.la to avoid any name collision
with flux-core's libjobspec.