resource: Update prune filters even for a jobspec without filter types
E.g., Given that --prune-filters=ALL:core and the jobsepc is

resources:
    - type: cluster
      count: 1
      with:
        - type: rack
          count: 1
          with:
            - type: slot
              count: 2
              label: default
              with:
                - type: node
                  count: 1

Since there is no change in the core-type resource, the old
scheme simply won't update the subtree planner
(so called prune filter).

However, this can lead to a side effect in that you can't use
this planner to retrieve all of the state changing
scheduled point.

The most notable problem arises with respect to the filter
at the root (e.g., cluster) vertex.

Within the dfu_traverser_t::schedule method, we use
planner_multi_avail_time_first and _next to quickly
traverse the resource state changing scheduled points.
In the worse case, if the previous jobspecs had not included
the type of resources specified in the prune filter,
this traversal will simply lead to a failure
-- finding no scheduled point!.