resource: Maintain at least one prune filter as default
Problem: With no prune filter installed, the schedule method
in dfu.cpp cannot correctly traverse the available scheduled
point for reservation.

If an argument to --prune-filters is given, we now append
these additional filter types to the list instead of
replacing the list.

Redundant types are resolved in the set_pruning_type method
by the use of std::set as the value field of the std::map.