plugin: add mod_name symbol to modified_fair_tree plugin
Problem: the plugin priority_mod_fair_tree.so does not export the
`mod_name` symbol as required for Flux Extension Modules as specified
in RFC 5. This causes an error whenever the flux-module utility
dlopen()'s this module to look for its name.

Add the `mod_name` symbol via the MOD_NAME() macro as required by
the spec, and ensure the symbol is exported in the DSO by adding
it to the -export-symbol-regex linker flag.

Fixes #433