qmanager: Add exception callback support
Exception callback enables a pending job (a job
with a pending alloc request) to be removed from
the scheduler queue when an exception is raised.
Running jobs that received an exception will still
be handled when corresponding free requests are
received.
Check if the target job that received an exception
with severity=0 is in the pending queue and if so
remove it from the system; otherwise, no-op.