resource: Add satisfiability checking match operation
Add MATCH_ALLOCATE_W_SATISFIABILITY.

This operation attempts to allocate first.
And if succeeds, it returns the matching info as before.

If fails, however, it sets the scheduled time to a point
as late as possible within the time box of the planner
and try the match.
Then, if it can't still find matching resources
at that scheduled point, the jobspec is deemed
unsatisfiable.
This satisfiability query must not change the visible
state of the graph data store. To support this,
overload the update method that only cleans up
the state changes created by the satisfiability
checking walk.

dfu_impl_t::schedule() (and therefore also run())
returns this information with -1 return code
with errno=ENODEV.
This is different than EBUSY, which is set when
dfu_impl_t::schedule() (and therefore also run())
cannot find the matching resources
at a given scheduled point (e.g., now).

Cleaned up how various errnos are propagated
in dfu_impl.cpp to simplify how errno=ENODEV
and errno=EBUSY are determined at dfu_impl_t::schedule().

Update in-line documentation with error numbers
for the public dfu_traverser_t::run() method.