resource: Add satisfiability checking match operation
Add MATCH_ALLOCATE_W_SATISFIABILITY.

This operation attempts to allocate first.
And if succeeds, it returns the matching info as before.

If fails, however, it sets the scheduled time to a point
as late as possible within the time box of the planner
and try the match.
Then, if it can't still find matching resources
at that scheduled point, the jobspec is deemed
unsatisfiable.
This satisfiability query must not change the visible
state of the graph data store. To support this,
overload the update method that only cleans up
the state changes created by the satisfiability
checking wakk.

select () returns this information with -1 return code
with errno set to ENODEV. This is different than EBUSY,
which is set when select () cannot find the matching
resources at a given scheduled point (e.g., NOW).

Add inline doc with those error numbers for the public
run() method.