resource: Use gettimeofday() as base time for plan-based scheduling
Problem: We use time 0 as base time for allocate
or allocate_orelse_reserve. While it was handy to test one
schedule loop within our resource infrastructure, this
doesn't allow for correct implementations for plan-based
scheduling algorithms including EAST and CONSERVATIVE.
For these algorithms, we need to move forward the base
time as the wallclock time moves forward.

Take the current time via gettimeofday() upon receiving each
match-allocate request within resource. If the matched
time is equal to the current time, we deem this case to
be "allocated." If the matched time is greater than the
current time, we deem this to be "reserved," instead.